refactor(path): 길찾기 API 의존성 감소를 위한 Mapper 패턴 추가 및 DTO 위치 이동 - intrefaces로 API DTO 이동 - 도메인 DTO 추가 - API 및 도메인 변환 담당 Mapper 패턴 추가 - 영향 범위 코드 리팩토링
