+++ backend/src/main/java/groom/backend/domain/path/service/spec/PathServiceImpl.java
@@ -3,7 +3,7 @@ package groom.backend.domain.path.service.spec;
 import groom.backend.domain.path.dto.request.PathFindRequest;
 import groom.backend.domain.path.dto.response.PathFindResponse;
 import groom.backend.domain.path.service.impl.PathService;
-import groom.backend.interfaces.TmapApiClient;
+import groom.backend.interfaces.tmap.TmapApiClient;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
