+++ backend/src/main/java/groom/backend/interfaces/tmap/dto/response/TmapPathFindResponse.java
@@ -1,4 +1,4 @@
-package groom.backend.domain.path.dto.response;
+package groom.backend.interfaces.tmap.dto.response;
 
 import groom.backend.domain.path.vo.PathNode;
 import groom.backend.domain.path.vo.PathSummary;
@@ -11,7 +11,7 @@ import java.util.List;
 @ToString
 @NoArgsConstructor
 @AllArgsConstructor
-public class PathFindResponse {
-   private PathSummary pathSummary;
-   private List<PathNode> pathNodeList;
+public class TmapPathFindResponse {
+  private PathSummary pathSummary;
+  private List<PathNode> pathNodeList;
 }
