+++ backend/src/test/java/groom/backend/path/TmapApiClientTest.java
@@ -2,7 +2,9 @@ package groom.backend.path;
 
 import groom.backend.domain.path.dto.request.PathFindRequest;
 import groom.backend.domain.path.dto.response.PathFindResponse;
-import groom.backend.interfaces.TmapApiClient;
+import groom.backend.interfaces.tmap.TmapApiClient;
+import groom.backend.interfaces.tmap.dto.response.TmapPathFindResponse;
+import groom.backend.interfaces.tmap.mapper.TmapToPathMapper;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -25,7 +27,7 @@ class TmapApiClientTest {
             "따흐앙"
     );
 
-    PathFindResponse res = tmapApiClient.tmapApiPathFind(req);
+    TmapPathFindResponse res = tmapApiClient.tmapApiPathFind(TmapToPathMapper.toTmapPathFindRequestDto(req));
     Assertions.assertNotNull(res);
     System.out.println(res);
   }
